# server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API to query USD current value",
        "title": "Server - GO Expert - Activity",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Diego Cordeiro",
            "url": "https://github.com/DiegoJCordeiro/client-server-activity-go",
            "email": "diegocordeiro.contatos@gmail.com"
        },
        "license": {
            "name": "Diego Cordeiro License",
            "url": "https://github.com/DiegoJCordeiro/client-server-activity-go/blob/main/LICENSE"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/quotation": {
            "get": {
                "description": "This endpoint is used to Query a quotation.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Quotation"
                ],
                "summary": "Query a quotation",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.QuotationOutputUseCaseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDTO"
                        }
                    }
                }
            },
            "put": {
                "description": "This endpoint is used to Update a quotation.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Quotation"
                ],
                "summary": "Update a quotation",
                "parameters": [
                    {
                        "description": "Quotation Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.QuotationInputHandlerUpdateDTO"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content",
                        "schema": {
                            "$ref": "#/definitions/dto.QuotationOutputUseCaseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDTO"
                        }
                    }
                }
            },
            "post": {
                "description": "This endpoint is used to Query a quotation.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Quotation"
                ],
                "summary": "Create a quotation",
                "parameters": [
                    {
                        "description": "Quotation Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.QuotationInputHandlerCreateDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.QuotationOutputUseCaseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDTO"
                        }
                    }
                }
            },
            "delete": {
                "description": "This endpoint is used to Delete a quotation.",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Quotation"
                ],
                "summary": "Delete a quotation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id quotation",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.QuotationOutputUseCaseDTO"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorDTO"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.ErrorDTO": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
        "dto.QuotationInputHandlerCreateDTO": {
            "type": "object",
            "properties": {
                "ask": {
                    "type": "string"
                },
                "bid": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "dto.QuotationInputHandlerUpdateDTO": {
            "type": "object",
            "properties": {
                "ask": {
                    "type": "string"
                },
                "bid": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "dto.QuotationOutputUseCaseDTO": {
            "type": "object",
            "properties": {
                "ask": {
                    "type": "string"
                },
                "bid": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        }
    }
}